requirements.txt
streamlit>=1.35.0
requests>=2.31.0
langchain-community
langchain
beautifulsoup4>=4.12.0
langchain-text-splitters>=0.2.0
sentence-transformers>=3.0.0
sentence-transformers[cross-encoder]>=3.0.0
torchvision
chromadb>=0.5.0
groq>=0.9.0
python-dotenv>=1.0.0
